feat: update action name and author in action.yml
This commit updates the action name and author in the action.yml file. The name is changed from 'Deploy to Vercel Action' to 'Vercel Action'. The author is changed from 'BetaHuhn' to 'EvanNotFound'.

docs: add new features to README.md

This commit adds information about new features in the README.md file. The new features are improvements from the version of [deploy-to-vercel-action](https://github.com/BetaHuhn/deploy-to-vercel-action). Now, the Vercel branch name is the same as the GitHub branch name.